+++ lib/models/controller_class.dart
@@ -1,0 +1,11 @@
+import 'package:meta/meta.dart';
+
+class ControllerClass {
+  final String tableName;
+  final String controller;
+
+  ControllerClass({
+    @required this.tableName,
+    @required this.controller,
+  });
+}