+++ CHANGELOG.md
@@ -1,3 +1,10 @@
-## 1.0.0
+## 0.0.2
 
-- Initial version, created by Stagehand
+- Added option to create controller
+- Controllers complete at basic level  
+- config:outputPath changed as entityPath
+
+
+## 0.0.1
+
+- ManagedObject creation completed